backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.4
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
faker==22.0.0
spacy==3.7.2
openai==1.6.1
openai-whisper==20231117
ffmpeg-python==0.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
imageio-ffmpeg==0.4.9
mediapipe==0.10.9
supabase==2.9.0
pydantic==2.5.3
pypdf==4.0.1
python-docx==1.1.0
passlib==1.7.4
argon2-cffi==23.1.0
reportlab==4.0.9
pdfplumber==0.10.3
pymupdf==1.23.8
ultralytics==8.0.196
